move datalist package into internal tree to not expose the API
